LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/processing-model-2/window-onerror-with-cross-frame-event-listeners-5-expected.txt

CONSOLE MESSAGE: Error: PASS


PASS window.onerror listener reports the exception in global object of its callback
